Перенёс всю логику в перезентер
